Refactor confirm dialog provider to use ZetkinConfirmDialog
