+++ playwright/tests/organize/people/views/delete-view.spec.ts
@@ -9,7 +9,7 @@ import KPD from '../../../../mockData/orgs/KPD';
 const deleteView = async (page: Page) => {
     await page.click('data-testid=EllipsisMenu-menuActivator');
     await page.click(`data-testid=EllipsisMenu-item-delete-view`);
-    await page.click('button > :text("Submit")');
+    await page.click('button > :text("Confirm")');
 };
 
 const expectDeleteViewError = async (page: Page) => {
