+++ src/locale/misc/en.yml
@@ -18,10 +18,7 @@ components:
             noEmpty: This cannot be empty
             edit: Click to edit
             save: Press 'enter' to save
-    ConfirmDialog:
-        defaultTitle: Confirm
-        defaultWarningText: Are you sure you want to do this action? It can not be undone.
-    confirm:
-        buttons:
-            submit: Confirm
-            cancel: Cancel
+ConfirmDialog:
+    defaultTitle: Confirm
+    defaultWarningText: Are you sure you want to do this action? It can not be undone.
+    button: Confirm
